test_verification_bot/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
supabase>=2.3.0
python-dotenv>=1.0.0
